Add class Comment to hold url and transform to json
