+++ portfolio/src/main/java/com/google/sps/servlets/Comment.java
@@ -1,0 +1,11 @@
+package com.google.sps.servlets;
+
+class Comment {
+    String message;
+    String imageUrl;
+
+    Comment(String message, String imageUrl) {
+        this.message = message;
+        this.imageUrl = imageUrl;
+    }
+}
